+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [Unreleased](https://github.com/convml/convml-data/tree/HEADER)
+
+[Full Changelog](https://github.com/convml/convml-data/compare/v0.2.1...)
+
+*enhancements*
+
+- Keep internal list of time for which CERES (SatCORPS) data is known to be
+  missing so that we don't request these files.
+  [\#21](https://github.com/convml/convml-data/pull/21)
+
+*bugfixes*
+
+- Fix creation of tile images from "multichannel" sources that arose from using
+  `matplotlib` to create the tile image, now images are created directly using
+  `PIL`.
+  [\#22](https://github.com/convml/convml-data/pull/22)
+
 
 ## [v0.2.1](https://github.com/convml/convml-data/tree/v0.2.1)